VS-5: Proper update of variant-bearing patients list
Variant documents were only holding on to one patient (callsetID) at a time, preventing patients in a common database from sharing a variant and being matched by it. This bug was occurring because the list meant to contain *all* pertinent callsetIDs was being re-set every time an exome containing that variant was uploaded. Bug was fixed with (a) the initialization of a cloneable callsetID field, (b) the creation of method addMultiFieldValue(), which clones the list of callsetIDs and appends to it (rather than resetting it.)